Fixed VS2010 compilation.
git-svn-id: http://www.bci2000.org/svn/trunk@4653 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/shared/types/GenericSignal.cpp
@@ -145,7 +145,7 @@ GenericSignal::SetProperties( const SignalProperties& inSp )
 ostream&
 GenericSignal::WriteToStream( ostream& os ) const
 {
-  string indent( os.width(), ' ' );
+  string indent( static_cast<size_t>( os.width() ), ' ' );
 
   os << '\n' << indent << "SignalProperties { ";
   mProperties.WriteToStream( os );
